# rust/broker/tests/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ConfigResponse": {
        "description": "`GET /api/config` body: feature flags the OWUI terminal UI connection-gates on.",
        "properties": {
          "features": {
            "$ref": "#/components/schemas/Features"
          }
        },
        "required": [
          "features"
        ],
        "type": "object"
      },
...
      "CreateSandboxRequest": {
        "description": "`POST /api/sandboxes` body: instantiate a Sandbox from a template.",
        "properties": {
          "name": {
            "description": "Desired `Sandbox` name (DNS-1123 label).",
            "type": "string"
          },
          "profile": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Profile",
                "description": "Persistence profile; defaults to `BROKER_DEFAULT_PROFILE` when omitted."
              }
            ]
          },
          "sessionId": {
            "description": "Owning session/chat id (recorded as the `broker-session` annotation).",
            "type": [
              "string",
              "null"
            ]
          },
          "templateName": {
            "description": "Name of the backing `SandboxTemplate` to clone the pod blueprint from.",
            "type": "string"
          },
          "userId": {
            "description": "Owning user id (recorded as the `broker-user` annotation).",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "templateName",
          "name"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Error body both control planes return: `{\"detail\": \"...\"}` (the\n`{ \"detail\": <string> }` shape — byte-for-byte parity, D11). Used as the\nerror response `body` across the broker and runtime `OpenAPI` surfaces.",
        "properties": {
          "detail": {
            "description": "Human-readable error detail.",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
      "Features": {
        "description": "Feature-flag set returned inside `ConfigResponse`.",
        "properties": {
          "notebooks": {
            "type": "boolean"
          },
          "system": {
            "type": "boolean"
          },
          "terminal": {
            "type": "boolean"
          }
        },
        "required": [
          "terminal",
          "notebooks",
          "system"
        ],
        "type": "object"
      },
...
      "SandboxObject": {
        "description": "`OpenAPI` schema for a `Sandbox` object on the wire (`apiVersion`/`kind`/`metadata`/\n`spec`/`status`). The CRUD handlers return the kube-generated [`shared::Sandbox`] (an\nidentical shape); this broker-local schema surfaces the typed spec/status in the\ngenerated document (issue #75).",
        "properties": {
          "apiVersion": {
            "description": "`agents.x-k8s.io/v1beta1`.",
            "type": "string"
          },
          "kind": {
            "description": "`Sandbox`.",
            "type": "string"
          },
          "metadata": {
            "description": "Kubernetes `ObjectMeta` (name, namespace, labels, annotations, …)."
          },
          "spec": {
            "$ref": "#/components/schemas/SandboxSpec",
            "description": "Sandbox spec (template, profile, operating mode, …)."
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SandboxStatus",
                "description": "Last-observed status (conditions, pod IPs, …); absent until reconciled."
              }
            ]
          }
        },
        "required": [
          "apiVersion",
          "kind",
          "metadata",
          "spec"
        ],
        "type": "object"
      },
      "SandboxSpec": {
        "description": "Spec of a [`Sandbox`] — a running tenant sandbox managed by the broker.\n\n`operatingMode` / `shutdownPolicy` are typed because the broker parks a\nsandbox (`Suspended`) and resumes it (`Running`) and relies on\n`shutdownPolicy: Retain` to keep the object across a park. `podTemplate` is\nthe opaque per-instance pod blueprint cloned from the backing\n[`SandboxTemplate`] (see the module docs for why it is untyped).",
        "properties": {
          "operatingMode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OperatingMode",
                "description": "Lifecycle mode the upstream controller honours. `Running` (default) keeps\nthe pod scheduled; `Suspended` parks it (pod deleted, object retained) so\na later resume reuses the same identity + per-session Secret."
              }
            ]
          },
          "podTemplate": {
            "description": "Opaque per-instance pod blueprint, cloned from the backing\n[`SandboxTemplate`] and shuffled by the broker (volumes, labels)."
          },
          "shutdownPolicy": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ShutdownPolicy",
                "description": "Whether the `Sandbox` object is retained after its pod exits. The broker\nsets `Retain` so a parked sandbox survives to be resumed rather than\nrecreated."
              }
            ]
          },
          "templateName": {
            "description": "Name of the [`SandboxTemplate`] backing this sandbox.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "SandboxStatus": {
        "description": "Status of a [`Sandbox`] — the subset of fields the broker reads (readiness +\npod IP) plus the common scalars it forwards. Authored here so readiness can\nbe computed in typed Rust rather than by poking raw JSON.",
        "properties": {
          "conditions": {
            "description": "Status conditions; a `Ready`/`True` entry means the sandbox is serving.",
            "items": {
              "$ref": "#/components/schemas/SandboxCondition"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "message": {
            "description": "Human-readable status message, if any.",
            "type": [
              "string",
              "null"
            ]
          },
          "phase": {
            "description": "High-level lifecycle phase (`Running`, `Suspended`, ...).",
            "type": [
              "string",
              "null"
            ]
          },
          "podIPs": {
            "description": "Pod IPs assigned to the sandbox; `[0]` is the address the broker proxies to.",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "ready": {
            "description": "Convenience readiness flag, when the controller surfaces it.",
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "StatusResponse": {
        "description": "`GET /api/status` body: static operator telemetry.",
        "properties": {
          "active_pods": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "max_pods": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "pods": {
            "items": {},
            "type": "array"
          }
        },
        "required": [
          "active_pods",
          "max_pods",
          "pods"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "brokerBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "the open-websandbox contributors"
    },
    "description": "Open WebUI front door + Kubernetes Sandbox lifecycle. The runtime tool surface (/execute, /files/*, /snapshot, /restore, /api/terminals*) is proxied to the resolved sandbox pod and documented here as part of the OWUI-facing contract. The catch-all reverse proxy (/{*path}) and the /api/terminals/{id} WebSocket upgrade are opaque passthroughs omitted from this document (issue #75 Q5).",
    "license": {
      "identifier": "AGPL-3.0-only",
      "name": "AGPL-3.0-only"
    },
    "title": "open-websandbox broker",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/config": {
      "get": {
        "operationId": "api_config",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            },
            "description": "Feature discovery flags"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing/invalid shared Bearer"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Shared secret not configured (fail-closed)"
          }
        },
        "security": [
          {
            "brokerBearer": []
          }
        ],
        "summary": "`GET /api/config` — feature discovery (terminal UI connection gate).",
        "tags": [
          "broker"
        ]
      }
    },
    "/api/sandboxes": {
      "get": {
        "operationId": "list_sandboxes",
        "parameters": [
          {
            "description": "Kubernetes label selector (URL `labelSelector`) to filter listed sandboxes.",
            "in": "path",
            "name": "labelSelector",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SandboxObject"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Broker-owned sandboxes"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "brokerBearer": []
          }
        ],
        "summary": "`GET /api/sandboxes` — list broker-owned sandboxes.",
        "tags": [
          "sandboxes"
        ]
      },
      "post": {
        "operationId": "create_sandbox",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSandboxRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxObject"
                }
              }
            },
            "description": "Already existed (idempotent create)"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxObject"
                }
              }
            },
            "description": "Sandbox created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing/invalid shared Bearer"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Template not found"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Apiserver rejected the call"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Shared secret not configured"
          }
        },
        "security": [
          {
            "brokerBearer": []
          }
        ],
        "summary": "`POST /api/sandboxes` — create a `Sandbox` from a `SandboxTemplate` (C-1\nfoundation; C-2's resolve-on-request reuses this path). 201 on create, 200\nwith the existing object on a 409 — idempotent (409 → update existing).",
        "tags": [
          "sandboxes"
        ]
      }
    },
    "/api/sandboxes/{name}": {
      "delete": {
        "operationId": "delete_sandbox",
        "parameters": [
          {
            "description": "Sandbox name",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted (404-tolerant → 204)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "brokerBearer": []
          }
        ],
        "summary": "`DELETE /api/sandboxes/{name}` — delete a sandbox (404-tolerant → 204).",
        "tags": [
          "sandboxes"
        ]
      },
      "get": {
        "operationId": "get_sandbox",
        "parameters": [
          {
            "description": "Sandbox name (DNS-1123 label)",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxObject"
                }
              }
            },
            "description": "Sandbox (status included)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sandbox not found"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "brokerBearer": []
          }
        ],
        "summary": "`GET /api/sandboxes/{name}` — get a single sandbox (status included).",
        "tags": [
          "sandboxes"
        ]
      }
    },
    "/api/status": {
      "get": {
        "operationId": "api_status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Operator telemetry (static)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "brokerBearer": []
          }
        ],
        "summary": "`GET /api/status` — operator telemetry (static until C-2/C-4 make it real).",
        "tags": [
          "broker"
        ]
      }
    },
...
    "/snapshot": {
      "get": {
        "operationId": "snapshot",
        "responses": {
          "200": {
            "description": "Workspace as a zstd-compressed tar stream (application/zstd)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing/invalid per-session Bearer"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Workspace exceeds MAX_WORKSPACE_BYTES"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Pipeline failure"
          }
        },
        "security": [
          {
            "brokerBearer": []
          }
        ],
        "summary": "`GET /snapshot` — stream the workspace as a zstd-compressed tar.",
        "tags": [
          "snapshot"
        ]
      }
    },
...
  },
...
}